pyproject.toml
[tool.poetry]
name = "ask_hn_digest"
version = "0.1.0"
description = "Get Summaries of the most engaging discussions on Hacker News"
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django-allauth = {extras = ["socialaccount"], version = "^64.0.0"}
python-webpack-boilerplate = "^1.0.0"
django-widget-tweaks = "^1.4.12"
mkdocs = "^1.4.2"
mkdocstrings = "^0.20.0"
mkdocs-material = "^9.0.12"
django = "^5.0.4"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
ipython = "^8.27.0"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
django-q2 = "^1.7.2"
whitenoise = "^6.7.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
structlog = "^24.4.0"
django-structlog = "^8.1.0"
markdown = "^3.7"
sentry-sdk = {extras = ["django"], version = "^2.14.0"}
structlog-sentry = "^2.2.1"
gunicorn = "^23.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
redis = "^5.0.8"
django-anymail = {extras = ["mailgun"], version = "^12.0"}
posthog = "^3.6.6"
django-ninja = "^1.3.0"
django-mjml = "^1.3"
pydantic-ai = "^0.0.43"
logfire = "^3.6.4"
pydantic = {extras = ["email"], version = "^2.11.4"}
[tool.poetry.dev-dependencies]


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.1"
pylint-django = "^2.5.3"
pre-commit = "^3.2.1"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.23"

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.djlint]
profile="django"
ignore = "H031,H006,H023,H021,H011,T002"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
